% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnrichMotifPair.R
\name{findEnrichMotifPair}
\alias{findEnrichMotifPair}
\title{Finds enriched motifs and motif pairs in given
genomic regions.}
\usage{
findEnrichMotifPair(
  target_data,
  background_data = NULL,
  genome_ver = "hg38",
  scramble_data = TRUE,
  motif_database = "JASPAR"
)
}
\arguments{
\item{target_data}{A data frame containing peaks coordinates with chr, start,
and end as first three columns}

\item{background_data}{Same as "target_data"}

\item{genome_ver}{Genome version, it should be either "hg19" or "hg38"}

\item{scramble_data}{A logical value (TRUE/FALSE) whether background data
need to be generated by scrambling target_data, by default FALSE}

\item{motif_database}{A character representing motif database to choose from,
it should be one of c("JASPAR", "ENCODE", "CISBP", "HOMER") default = "JASPAR"}
}
\value{
A list with two data frames containing enriched motifs and motif
pairs with P-values, fold-enrichment and annotations
}
\description{
Finds enriched motifs and motif pairs in given
genomic regions.
}
\examples{
\dontrun{
enrichMotif(target_data, background_data = NULL,
genome_ver = "hg38", scramble_data = FALSE)
}
}
