% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnrichMotifPair_Jolma2015.R
\name{findEnrichMotifPair_Jolma2015}
\alias{findEnrichMotifPair_Jolma2015}
\title{Finds enriched motif pairs in given a
genomic regions.}
\usage{
findEnrichMotifPair_Jolma2015(
  target_data,
  background_data = NULL,
  genome_ver = "hg38",
  scramble_data = TRUE,
  Pvalue_computation = "hyper",
  Pvalue_threshold = 0.05,
  Pvalue_adjust_method = "BH"
)
}
\arguments{
\item{target_data}{A data frame containing peaks coordinates with chr, start,
and end as first three columns}

\item{background_data}{Same as "target_data"}

\item{genome_ver}{Genome version, it should be either "hg19" or "hg38"}

\item{scramble_data}{A logical value (TRUE/FALSE) whether background data
need to be generated by scrambling target_data, by default FALSE}

\item{Pvalue_computation}{A character representing P value computations are
either based on binomial or hyper geometric distributions, it should be one
of c("binom", "hyper"), default = "binom"}

\item{Pvalue_threshold}{A numeric value for filtering the enriched motifs and
motif pairs}

\item{Pvalue_adjust_method}{A character representing method for adjustig the
P values for multple comparisons, it should be one or more of
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")
default = "BH"}
}
\value{
A list with two data frames containing enriched motifs and motif
pairs with P-values, fold-enrichment and annotations
}
\description{
Finds enriched motif pairs in given a
genomic regions.
}
\examples{
\dontrun{
findEnrichMotifPair_Jolma2015(target_data, background_data = NULL,
genome_ver = "hg38", scramble_data = TRUE)
}
}
