% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBackgroundSeqs.R
\name{generateBackgroundSeqs}
\alias{generateBackgroundSeqs}
\title{Finds GC, length and chromosome matched random genomic sequences}
\usage{
generateBackgroundSeqs(
  input_data,
  genome_ver = "hg38",
  xfold = 1,
  batchsize = 30000,
  nMaxTrials = 50
)
}
\arguments{
\item{input_data}{A data frame containing peaks coordinates with chr, start,
and end as first three columns}

\item{genome_ver}{Genome version, it should be either "hg19" or "hg38"}

\item{xfold}{An integer, number of times the backgrouns seqs should be
generated}

\item{batchsize}{An integer, number of random sequences tested in each trial}

\item{nMaxTrials}{An integer, number of maximum trials}
}
\value{
A GRanges object with coordinates
}
\description{
Finds GC, length and chromosome matched random genomic sequences
}
\examples{
\dontrun{
generateBackgroundSeqs(input_data, genome_ver = "hg38")
}
}
