% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScrambleSeqs.R
\name{generateScrambleSeqs}
\alias{generateScrambleSeqs}
\title{Generates scrambled sequences from genomic regions}
\usage{
generateScrambleSeqs(input_data, genome_ver = "hg38")
}
\arguments{
\item{input_data}{A data frame containing peaks coordinates with chr, start,
and end as first three columns}

\item{genome_ver}{Genome version, it should be either "hg19" or "hg38"}
}
\value{
A DNAStringSet instance with scrambled seqs
}
\description{
Generates scrambled sequences from genomic regions
}
\examples{
\dontrun{
generateScrambleSeqs(input_data, genome_ver = "hg38")
}
}
