% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_data.R
\docType{data}
\name{peak_data}
\alias{peak_data}
\title{Enhancer peak coordinates}
\format{A data frame with 13051 rows and 4 variables:
\describe{
  \item{chr}{chromosome in hg38}
  \item{start}{start in hg38}
  \item{end}{end in hg38}
  \item{peakid}{peak identifier}
  ...
}}
\usage{
peak_data
}
\description{
A dataset containing the enhancer peak coordinates for total CD4+ cell from
BLUEPRINT consortium
}
\keyword{datasets}
