% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEnrichPair}
\alias{plotEnrichPair}
\title{Plots the enrichment values of selected TFs.}
\usage{
plotEnrichPair(enrich_pairs, tfs)
}
\arguments{
\item{tfs}{TFs for which to plot enrichment}

\item{enrich_pair}{motif_pair_enrich output dataframe from findEnrichMotifPair}

\item{metric}{Value to plot; either pval or logFC}
}
\value{
a ggplot2 plot object
}
\description{
Plots the enrichment values of selected TFs.
}
\examples{
\dontrun{
motif_pairs_df <- results$motif_pair_enrich
plotEnrichment(enrich_motifs = motif_pairs_df, tfs = c("SOX2", "NANOG", "MYC", "POU5F1"),
metric = "pval")
}
}
