% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plots network of TF enrichment interactions.}
\usage{
plotNetwork(
  enrich_pairs,
  TF_name,
  color_TF = "#70d9e0",
  color_bind_TF = "#e841da"
)
}
\arguments{
\item{color_TF}{Color of input TF}

\item{color_bind_TF}{Color of binding partner TFs}

\item{enrich_pair}{motif_pair_enrich output dataframe from findEnrichMotifPair}

\item{tfs}{TF for which to plot enrichment network}
}
\value{
a ggplot2 plot object
}
\description{
Plots network of TF enrichment interactions.
}
\examples{
\dontrun{
motif_pairs_df <- results$motif_pair_enrich
plotNetwork(enrich_pairs, TF_name = "SP8")
}
}
