% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDuplicateTFPairs.R
\name{removeDuplicateTFPairs}
\alias{removeDuplicateTFPairs}
\title{Finds and removes duplicate motif pairs from the output of
"findEnrichMotifPair" or "findEnrichMotifPairAll"}
\usage{
removeDuplicateTFPairs(data = data)
}
\arguments{
\item{data}{A data frame, the output of "findEnrichMotifPair" or "findEnrichMotifPairAll"}
}
\value{
A data frame same as input containing unique (non-redundant) motif pairs with
P-values, fold-enrichment and annotations
}
\description{
Finds and removes duplicate motif pairs from the output of
"findEnrichMotifPair" or "findEnrichMotifPairAll"
}
\examples{
\dontrun{
removeDuplicateTFPairs(data)
}
}
