% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runEnrichMotifPair}
\alias{runEnrichMotifPair}
\title{Opens interactive shiny to run enrichmotifpairR.}
\usage{
runEnrichMotifPair(...)
}
\description{
Opens interactive shiny to run enrichmotifpairR.
}
\examples{
\dontrun{
runEnrichMotifPair()
}
}
